Backend_Django/pyproject.toml
[tool.poetry]
name = "Backend_Django"
version = "0.1.0"
description = ""
authors = ["Rasoul Mesghali <rasoul.mesghali@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.2"
djangorestframework = "^3.12.1"
dj_database_url = "^0.5.0"
django-dotenv = "^1.4.2"
psycopg2 = "^2.8.6"
coloredlogs = "^14.0"
requests = "^2.24.0"
django-cors-headers = "^3.5.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^5.0.0"}
django-log-request-id = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
